requirements.txt
streamlit
google-genai
pandas
numpy
matplotlib
scipy